pyproject.toml
[project]
name = "mma-meeting"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "gae", email = "gaedev22@gmail.com" }
]
dependencies = [
    "bs4>=0.0.2",
    "beautifulsoup4>=4.12.3",
    "flask>=3.0.3",
    "google-auth>=2.35.0",
    "gspread>=6.1.2",
    "pyparsing>=3.1.4",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "flask-ldap3-login>=1.0.2",
    "flask-wtf>=1.2.1",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.147.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mma_meeting"]
